pyproject.toml
[project]
name = "route-optimizer"
version = "0.1.0"
description = ""
authors = [
    {name = "Ryan Taylor",email = "rataylor93@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "openrouteservice (>=2.3.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]

[tool.poetry]
packages = [{include = "route_optimizer", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
